Auth/requirements.txt
alembic==1.7.7
apispec==5.2.1
async-timeout==4.0.2
attrs==21.4.0
Authlib==1.0.1
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
commonmark==0.9.1
cryptography==37.0.2
Deprecated==1.2.13
dominate==2.6.0
flasgger==0.9.5
Flask==2.1.1
Flask-Bootstrap4==4.0.2
Flask-JWT-Extended==4.3.1
Flask-Limiter==2.4.5.1
Flask-Migrate==3.1.0
flask-paginate==2022.1.8
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.46.1
grpcio-tools==1.46.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.5.1
limits==2.6.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
mistune==2.0.2
msgpack==1.0.3
oauthlib==3.2.0
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-exporter-jaeger-proto-grpc==1.10.0
opentelemetry-exporter-jaeger-thrift==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-instrumentation-wsgi==0.29b0
opentelemetry-sdk==1.10.0
opentelemetry-semantic-conventions==0.29b0
opentelemetry-util-http==0.29b0
packaging==21.3
pluggy==1.0.0
protobuf==3.20.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyotp==2.6.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.4.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.35
thrift==0.16.0
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
